+++ app/src/main/java/com/example/capstoneproject/activity/SignUpActivity.java
@@ -12,9 +12,8 @@ import androidx.appcompat.widget.AppCompatButton;
 import com.example.capstoneproject.R;
 import com.example.capstoneproject.data.auth.AuthService;
 import com.example.capstoneproject.data.auth.request.User;
-import com.example.capstoneproject.data.auth.response.LoginResponse;
-import com.example.capstoneproject.data.auth.response.DuplicateResponse;
-import com.example.capstoneproject.data.auth.response.SignUpResponse;
+import com.example.capstoneproject.data.auth.response.duplicateuid.DuplicateResponse;
+import com.example.capstoneproject.data.auth.response.signup.SignUpResponse;
 import com.example.capstoneproject.view.DuplicateView;
 import com.example.capstoneproject.view.SignUpView;
 import com.google.android.material.textfield.TextInputEditText;
