+++ app/src/main/java/com/example/capstoneproject/data/auth/response/duplicateuid/CheckUidResult.java
@@ -1,4 +1,4 @@
-package com.example.capstoneproject.data.auth.response.result;
+package com.example.capstoneproject.data.auth.response.duplicateuid;
 
 import com.google.gson.annotations.SerializedName;
