+++ app/src/main/java/com/example/capstoneproject/view/DuplicateView.java
@@ -1,6 +1,6 @@
 package com.example.capstoneproject.view;
 
-import com.example.capstoneproject.data.auth.response.DuplicateResponse;
+import com.example.capstoneproject.data.auth.response.duplicateuid.DuplicateResponse;
 
 public interface DuplicateView {
     void onCheckedSuccess();
