+++ app/src/main/java/com/example/capstoneproject/view/LoginView.java
@@ -1,7 +1,6 @@
 package com.example.capstoneproject.view;
 
-import com.example.capstoneproject.data.auth.response.LoginResponse;
-import com.example.capstoneproject.data.auth.response.result.LoginResult;
+import com.example.capstoneproject.data.auth.response.login.LoginResult;
 
 public interface LoginView {
     void onLoginSuccess(int code, LoginResult result);
