+++ app/src/main/java/com/example/capstoneproject/view/SignUpView.java
@@ -1,7 +1,6 @@
 package com.example.capstoneproject.view;
 
-import com.example.capstoneproject.data.auth.response.LoginResponse;
-import com.example.capstoneproject.data.auth.response.SignUpResponse;
+import com.example.capstoneproject.data.auth.response.signup.SignUpResponse;
 
 public interface SignUpView {
     void onSignUpSuccess();
